+++ apps/api/src/app.module.ts
@@ -6,6 +6,7 @@ import { validateEnv } from './config/env.validation';
 import { AuthModule } from './identity/auth/auth.module';
 import { HealthModule } from './health/health.module';
 import { IdentityModule } from './identity/identity.module';
+import { OrganisationModule } from './identity/organisation/organisation.module';
 import { PrismaModule } from './prisma/prisma.module';
 
 @Module({
@@ -19,6 +20,7 @@ import { PrismaModule } from './prisma/prisma.module';
     HealthModule,
     IdentityModule,
     AuthModule,
+    OrganisationModule,
   ],
 })
 export class AppModule {}